pyproject.toml
[project]
name = "platform-backend"
version = "0.1.0"
description = "Unified platform backend serving Socialwise and JusMonitorIA"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.133",
    "uvicorn[standard]>=0.41",
    "sse-starlette>=2.0",
    "httpx>=0.28",
    "pydantic[email]>=2.5",
    "pydantic-settings>=2.1",

    # Task Queue
    "taskiq>=0.12",
    "taskiq-redis>=1.2",

    # AI
    "langgraph>=1.0",
    "litellm>=1.81",
    "langchain>=1.2",
    "langchain-openai>=1.1",
    "langchain-anthropic>=1.3",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.31",
    "alembic>=1.13",
    "pgvector>=0.4",

    # Redis
    "redis>=7.2",

    # Auth & Security
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=5.0",
    "cryptography>=46.0",
    "python-multipart>=0.0.22",

    # Storage
    "boto3>=1.35",

    # Document Processing
    "Pillow>=11.0",
    "PyMuPDF>=1.25",
    "pdf2image>=1.17",
    "python-docx>=1.2",
    "pypdf>=5.0",
    "reportlab>=4.0",

    # Logging & Monitoring
    "structlog>=25.5",
    "prometheus-client>=0.24",

    # Utilities
    "python-dotenv>=1.0",
    "croniter>=6.0",
    "aiosmtplib>=5.1",
    "google-genai>=1.66",
]

[project.optional-dependencies]
# JusMonitorIA domain-specific deps
jusmonitoria = [
    "zeep>=4.2",
    "pyhanko>=0.25",
    "python-docx>=1.2",
    "weasyprint>=68.0",
    "pyzbar>=0.1.9",
    "pyotp>=2.9",
    "qrcode>=8.0",
    "unidecode>=1.3",
    "faker>=40.5",
    "requests>=2.32",
]

dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-cov>=7.0",
    "pytest-mock>=3.12",
    "ruff>=0.15",
    "mypy>=1.8",
    "black>=26.1",
    "isort>=8.0",
    "sqlacodegen>=3.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["platform_core", "domains", "scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
